+++ src/app/delivery/delivery.component.ts
@@ -1,4 +1,6 @@
 import { Component } from '@angular/core';
+import { Router, ActivatedRoute } from '@angular/router';
+import { SessionStorageService } from '../core/services/session-storage.service';
 
 @Component({
   selector: 'app-delivery',
@@ -6,5 +8,14 @@ import { Component } from '@angular/core';
   styleUrls: ['./delivery.component.css']
 })
 export class DeliveryComponent {
+  constructor(private sessionStorageService: SessionStorageService,private router: Router, private route: ActivatedRoute) { }
+  type!:string|null;
+
+  ngOnInit() {
+    this.type = this.sessionStorageService.getItem('type');
+    if(this.type != 'livreur'){
+      this.router.navigate([`/error-page`], { relativeTo: this.route });
+    }
+  }
 
 }