+++ src/app/sales/widget/users-info/users-info.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { userSalesModel } from 'src/app/core/models/user-sales.model';
 
 @Component({
   selector: 'app-users-info',
@@ -6,5 +7,15 @@ import { Component } from '@angular/core';
   styleUrls: ['./users-info.component.css']
 })
 export class UsersInfoComponent {
+  usersList:userSalesModel[] = [];
 
+  ngOnInit(): void {
+    this.usersList.push(
+      { mail: "amara.ahmed@gmail.com", type: "client"},
+      { mail: "amara2.ahmed@gmail.com", type: "restaurant"},
+      { mail: "amara3.ahmed@gmail.com", type: "livreur"},
+      { mail: "amara4.ahmed@gmail.com", type: "Développeur Tiers"},
+      { mail: "amara5.ahmed@gmail.com", type: "client"},
+    );
+  }
 }